Update camera.py
Fixed wrong data display in heatmap. It seems that the camera is starting-up the first measurement after the DDL is started. Since we restart the DDL every time with a new measurement we display the "start-up" signal instead of an actual measurement. I fixed it by introducing an extra block in the settings of camera.py, and return the second block